requirements.txt
# GitHub Trending - 依赖列表

# 基础依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
pydantic>=2.0.0
click>=8.1.0
lxml>=4.9.0

# AI 客户端（至少安装一个）
anthropic>=0.18.0      # Claude
openai>=1.0.0          # OpenAI / DeepSeek / Ollama 通用客户端

# 可视化（可选）
matplotlib>=3.7.0

# 数据处理（可选）
pandas>=2.0.0
